renamed files & fixed style
